update the scrooge-generated code for Scrooge 22.4.0
also moves the generated code to input/src_managed, so that the scripts
can differentiate between the generated and manually-written inputs
