+++ scalafix/input/src_managed/main/scala/example/thrift/SimpleService$FinagleService.scala
@@ -1,9 +1,9 @@
 /*rule = AddCatsTaglessInstances*/
 /**
  * Generated by Scrooge
- *   version: 22.3.0
- *   rev: 2f6ae2f0908271965ba99fd3643c2d35b1e69a10
- *   built at: 20220329-191327
+ *   version: 22.4.0
+ *   rev: d1de8299d70540617bc183010bb06ef82acb88df
+ *   built at: 20220419-215016
  */
 package example.thrift
