+++ scalafix/output/src_managed/main/scala/example/thrift/SimpleResponse.scala
@@ -1,9 +1,8 @@
-/*rule = AddCatsTaglessInstances*/
 /**
  * Generated by Scrooge
- *   version: 22.3.0
- *   rev: 2f6ae2f0908271965ba99fd3643c2d35b1e69a10
- *   built at: 20220329-191327
+ *   version: 22.4.0
+ *   rev: d1de8299d70540617bc183010bb06ef82acb88df
+ *   built at: 20220419-215016
  */
 package example.thrift
 
@@ -29,8 +28,8 @@ import scala.collection.mutable.Builder
 import scala.reflect.{ClassTag, classTag}
 
 
-object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with StructBuilderFactory[SimpleRequest] {
-  val Struct: TStruct = new TStruct("SimpleRequest")
+object SimpleResponse extends ValidatingThriftStructCodec3[SimpleResponse] with StructBuilderFactory[SimpleResponse] {
+  val Struct: TStruct = new TStruct("SimpleResponse")
   val IdField: TField = new TField("id", TType.STRING, 1)
   val IdFieldManifest: Manifest[String] = manifest[String]
 
@@ -60,22 +59,22 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
     classTag[String].asInstanceOf[ClassTag[_]]
   )
 
-  private[this] val structFields: Seq[ThriftStructField[SimpleRequest]] = Seq[ThriftStructField[SimpleRequest]](
-    new ThriftStructField[SimpleRequest](
+  private[this] val structFields: Seq[ThriftStructField[SimpleResponse]] = Seq[ThriftStructField[SimpleResponse]](
+    new ThriftStructField[SimpleResponse](
       IdField,
       _root_.scala.Some(IdFieldManifest),
-      classOf[SimpleRequest]) {
-        def getValue[R](struct: SimpleRequest): R = struct.id.asInstanceOf[R]
+      classOf[SimpleResponse]) {
+        def getValue[R](struct: SimpleResponse): R = struct.id.asInstanceOf[R]
     }
   )
 
-  override lazy val metaData: ThriftStructMetaData[SimpleRequest] =
+  override lazy val metaData: ThriftStructMetaData[SimpleResponse] =
     ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
 
   /**
    * Checks that all required fields are non-null.
    */
-  def validate(_item: SimpleRequest): Unit = {
+  def validate(_item: SimpleResponse): Unit = {
     if (_item.id eq null) throw new TProtocolException("Required field id cannot be null")
   }
 
@@ -83,7 +82,7 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
    * Checks that the struct is a valid as a new instance. If there are any missing required or
    * construction required fields, return a non-empty list.
    */
-  def validateNewInstance(item: SimpleRequest): scala.Seq[com.twitter.scrooge.validation.Issue] = {
+  def validateNewInstance(item: SimpleResponse): scala.Seq[com.twitter.scrooge.validation.Issue] = {
     val buf = scala.collection.mutable.ListBuffer.empty[com.twitter.scrooge.validation.Issue]
 
     if (item.id eq null)
@@ -96,18 +95,18 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
    * Validate that all validation annotations on the struct meet the criteria defined in the
    * corresponding [[com.twitter.scrooge.validation.ThriftConstraintValidator]].
    */
-  def validateInstanceValue(item: SimpleRequest): Set[com.twitter.scrooge.thrift_validation.ThriftValidationViolation] = {
+  def validateInstanceValue(item: SimpleResponse): Set[com.twitter.scrooge.thrift_validation.ThriftValidationViolation] = {
     val violations = scala.collection.mutable.Set.empty[com.twitter.scrooge.thrift_validation.ThriftValidationViolation]
     violations ++= validateFieldValue("id", item.id, fieldInfos.apply(0).fieldAnnotations, scala.None)
     violations.toSet
   }
 
-  def withoutPassthroughFields(original: SimpleRequest): SimpleRequest =
+  def withoutPassthroughFields(original: SimpleResponse): SimpleResponse =
     new Immutable(
       id = original.id
     )
 
-  lazy val unsafeEmpty: SimpleRequest = {
+  lazy val unsafeEmpty: SimpleResponse = {
     val id: String = "empty"
 
     new Immutable(
@@ -116,14 +115,14 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
     )
   }
 
-  def newBuilder(): StructBuilder[SimpleRequest] = new SimpleRequestStructBuilder(_root_.scala.None, fieldTypes)
+  def newBuilder(): StructBuilder[SimpleResponse] = new SimpleResponseStructBuilder(_root_.scala.None, fieldTypes)
 
-  override def encode(_item: SimpleRequest, _oproto: TProtocol): Unit = {
+  override def encode(_item: SimpleResponse, _oproto: TProtocol): Unit = {
     _item.write(_oproto)
   }
 
 
-  override def decode(_iprot: TProtocol): SimpleRequest = {
+  override def decode(_iprot: TProtocol): SimpleResponse = {
     if (_iprot.isInstanceOf[LazyTProtocol]) {
       decodeInternal(_iprot, true)
     } else {
@@ -131,11 +130,11 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
     }
   }
 
-  private[thrift] def eagerDecode(_iprot: TProtocol): SimpleRequest = {
+  private[thrift] def eagerDecode(_iprot: TProtocol): SimpleResponse = {
     decodeInternal(_iprot, false)
   }
 
-  private[this] def decodeInternal(_iprot: TProtocol, lazily: Boolean): SimpleRequest = {
+  private[this] def decodeInternal(_iprot: TProtocol, lazily: Boolean): SimpleResponse = {
     var idOffset: Int = -1
     var id: String = null
     var _got_id = false
@@ -167,7 +166,7 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
     } while (!_done)
     _iprot.readStructEnd()
 
-    if (!_got_id) _root_.com.twitter.scrooge.internal.TProtocols.throwMissingRequiredField("SimpleRequest", "id")
+    if (!_got_id) _root_.com.twitter.scrooge.internal.TProtocols.throwMissingRequiredField("SimpleResponse", "id")
 
     val _passthroughFieldsResult =
       if (_passthroughFields eq null) _root_.com.twitter.scrooge.internal.TProtocols.NoPassthroughFields
@@ -192,30 +191,30 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
 
   def apply(
     id: String
-  ): SimpleRequest =
+  ): SimpleResponse =
     new Immutable(
       id
     )
 
-  def unapply(_item: SimpleRequest): _root_.scala.Option[String] = _root_.scala.Some(_item.id)
+  def unapply(_item: SimpleResponse): _root_.scala.Option[String] = _root_.scala.Some(_item.id)
 
 
 
-  object Immutable extends ThriftStructCodec3[SimpleRequest] {
-    override def encode(_item: SimpleRequest, _oproto: TProtocol): Unit = { _item.write(_oproto) }
-    override def decode(_iprot: TProtocol): SimpleRequest = SimpleRequest.decode(_iprot)
-    override lazy val metaData: ThriftStructMetaData[SimpleRequest] = SimpleRequest.metaData
+  object Immutable extends ThriftStructCodec3[SimpleResponse] {
+    override def encode(_item: SimpleResponse, _oproto: TProtocol): Unit = { _item.write(_oproto) }
+    override def decode(_iprot: TProtocol): SimpleResponse = SimpleResponse.decode(_iprot)
+    override lazy val metaData: ThriftStructMetaData[SimpleResponse] = SimpleResponse.metaData
   }
 
   /**
-   * The default read-only implementation of SimpleRequest.  You typically should not need to
-   * directly reference this class; instead, use the SimpleRequest.apply method to construct
+   * The default read-only implementation of SimpleResponse.  You typically should not need to
+   * directly reference this class; instead, use the SimpleResponse.apply method to construct
    * new instances.
    */
   class Immutable(
       val id: String,
       override val _passthroughFields: immutable$Map[Short, TFieldBlob])
-    extends SimpleRequest {
+    extends SimpleResponse {
     def this(
       id: String
     ) = this(
@@ -235,7 +234,7 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
       _end_offset: Int,
       idOffset: Int,
       override val _passthroughFields: immutable$Map[Short, TFieldBlob])
-    extends SimpleRequest {
+    extends SimpleResponse {
 
     override def write(_oprot: TProtocol): Unit = {
       if (_oprot.isInstanceOf[LazyTProtocol]) {
@@ -269,16 +268,16 @@ object SimpleRequest extends ValidatingThriftStructCodec3[SimpleRequest] with St
 }
 
 /**
- * Prefer the companion object's [[example.thrift.SimpleRequest.apply]]
+ * Prefer the companion object's [[example.thrift.SimpleResponse.apply]]
  * for construction if you don't need to specify passthrough fields.
  */
-trait SimpleRequest
+trait SimpleResponse
   extends ThriftStruct
   with _root_.scala.Product1[String]
-  with ValidatingThriftStruct[SimpleRequest]
+  with ValidatingThriftStruct[SimpleResponse]
   with java.io.Serializable
 {
-  import SimpleRequest._
+  import SimpleResponse._
 
   def id: String
 
@@ -304,7 +303,7 @@ trait SimpleRequest
         case 1 =>
           if (id ne null) {
             _oprot.writeString(id)
-            _root_.scala.Some(SimpleRequest.IdField)
+            _root_.scala.Some(SimpleResponse.IdField)
           } else {
             _root_.scala.None
           }
@@ -331,7 +330,7 @@ trait SimpleRequest
    * is unknown and passthrough fields are enabled, then the blob will be stored in
    * _passthroughFields.
    */
-  def setField(_blob: TFieldBlob): SimpleRequest = {
+  def setField(_blob: TFieldBlob): SimpleResponse = {
     var id: String = this.id
     var _passthroughFields = this._passthroughFields
     val _iprot = _blob.read 
@@ -351,7 +350,7 @@ trait SimpleRequest
    * known, it is reverted to its default value; if the field is unknown, it is removed
    * from the passthroughFields map, if present.
    */
-  def unsetField(_fieldId: Short): SimpleRequest = {
+  def unsetField(_fieldId: Short): SimpleResponse = {
     var id: String = this.id
 
     _fieldId match {
@@ -370,11 +369,11 @@ trait SimpleRequest
    * known, it is reverted to its default value; if the field is unknown, it is removed
    * from the passthroughFields map, if present.
    */
-  def unsetId: SimpleRequest = unsetField(1)
+  def unsetId: SimpleResponse = unsetField(1)
 
 
   override def write(_oprot: TProtocol): Unit = {
-    SimpleRequest.validate(this)
+    SimpleResponse.validate(this)
     _oprot.writeStructBegin(Struct)
     if (id ne null) {
       _oprot.writeFieldBegin(IdField)
@@ -387,21 +386,21 @@ trait SimpleRequest
   def copy(
     id: String = this.id,
     _passthroughFields: immutable$Map[Short, TFieldBlob] = this._passthroughFields
-  ): SimpleRequest =
+  ): SimpleResponse =
     new Immutable(
       id,
       _passthroughFields
     )
 
-  override def canEqual(other: Any): Boolean = other.isInstanceOf[SimpleRequest]
+  override def canEqual(other: Any): Boolean = other.isInstanceOf[SimpleResponse]
 
-  private[this] def _equals(other: SimpleRequest): Boolean =
+  private[this] def _equals(other: SimpleResponse): Boolean =
       this.productArity == other.productArity &&
       this.productIterator.sameElements(other.productIterator) &&
       this._passthroughFields == other._passthroughFields
 
   override def equals(other: Any): Boolean =
-    canEqual(other) && _equals(other.asInstanceOf[SimpleRequest])
+    canEqual(other) && _equals(other.asInstanceOf[SimpleResponse])
 
   override def hashCode: Int = {
     _root_.scala.runtime.ScalaRunTime._hashCode(this)
@@ -409,29 +408,28 @@ trait SimpleRequest
 
   override def toString: String = _root_.scala.runtime.ScalaRunTime._toString(this)
 
-  override def productPrefix: String = "SimpleRequest"
+  override def productPrefix: String = "SimpleResponse"
 
-  def _codec: ValidatingThriftStructCodec3[SimpleRequest] = SimpleRequest
+  def _codec: ValidatingThriftStructCodec3[SimpleResponse] = SimpleResponse
 
-  def newBuilder(): StructBuilder[SimpleRequest] = new SimpleRequestStructBuilder(_root_.scala.Some(this), fieldTypes)
+  def newBuilder(): StructBuilder[SimpleResponse] = new SimpleResponseStructBuilder(_root_.scala.Some(this), fieldTypes)
 }
 
-private[thrift] class SimpleRequestStructBuilder(instance: _root_.scala.Option[SimpleRequest], fieldTypes: IndexedSeq[ClassTag[_]])
-    extends StructBuilder[SimpleRequest](fieldTypes) {
+private[thrift] class SimpleResponseStructBuilder(instance: _root_.scala.Option[SimpleResponse], fieldTypes: IndexedSeq[ClassTag[_]])
+    extends StructBuilder[SimpleResponse](fieldTypes) {
 
-  def build(): SimpleRequest = {
+  def build(): SimpleResponse = {
     val _fieldArray = fieldArray // shadow variable
     if (instance.isDefined) {
       val instanceValue = instance.get
-      SimpleRequest(
+      SimpleResponse(
         if (_fieldArray(0) == null) instanceValue.id else _fieldArray(0).asInstanceOf[String]
       )
     } else {
-      if (genericArrayOps(_fieldArray).contains(null)) throw new InvalidFieldsException(structBuildError("SimpleRequest"))
-      SimpleRequest(
+      if (genericArrayOps(_fieldArray).contains(null)) throw new InvalidFieldsException(structBuildError("SimpleResponse"))
+      SimpleResponse(
         _fieldArray(0).asInstanceOf[String]
       )
     }
   }
 }
-
