+++ scalafix/output/src_managed/main/scala/example/thrift/SimpleService.scala
@@ -1,9 +1,8 @@
-/*rule = AddCatsTaglessInstances*/
 /**
  * Generated by Scrooge
- *   version: 22.3.0
- *   rev: 2f6ae2f0908271965ba99fd3643c2d35b1e69a10
- *   built at: 20220329-191327
+ *   version: 22.4.0
+ *   rev: d1de8299d70540617bc183010bb06ef82acb88df
+ *   built at: 20220419-215016
  */
 package example.thrift
 
@@ -738,15 +737,24 @@ object SimpleService extends _root_.com.twitter.finagle.thrift.GeneratedThriftSe
   type makeRequest$result = MakeRequest.Result
 
 
-  trait MethodPerEndpoint extends _root_.com.twitter.finagle.thrift.ThriftService {
+  trait SimpleService[F[_]] extends _root_.com.twitter.finagle.thrift.ThriftService {
     
-    def makeRequest(request: example.thrift.SimpleRequest): Future[example.thrift.SimpleResponse]
+    def makeRequest(request: example.thrift.SimpleRequest): F[example.thrift.SimpleResponse]
     /**
      * Used to close the underlying `Service`.
      * Not a user-defined API.
      */
     def asClosable: _root_.com.twitter.util.Closable = _root_.com.twitter.util.Closable.nop
   }
+
+  object SimpleService {
+    implicit def SimpleServiceInReaderT[F[_]]: SimpleService[({type Λ[β0] = _root_.cats.data.ReaderT[F, SimpleService[F], β0]})#Λ] =
+      _root_.cats.tagless.Derive.readerT[SimpleService, F]
+
+    implicit val SimpleServiceFunctorK: _root_.cats.tagless.FunctorK[SimpleService] = _root_.cats.tagless.Derive.functorK[SimpleService]
+  }
+
+  trait MethodPerEndpoint extends SimpleService[Future]
 
   object MethodPerEndpoint {
